refactor: move tool outside of lib
