+++ packages/board_generator/tool/model/model.dart
