+++ packages/board_generator/tool/symmetrical_generation.dart
@@ -1,10 +1,13 @@
 import 'dart:io';
 
 import 'package:board_generator/board_generator.dart';
-import 'package:board_generator/tool/model/model.dart';
 import 'package:csv/csv.dart';
 
-void main({void Function(String) log = print}) {
+import 'model/model.dart';
+
+void main({
+  void Function(String) log = print,
+}) {
   final fileString = File('words.csv').readAsStringSync();
   final rows = const CsvToListConverter().convert(fileString);
 
@@ -16,25 +19,30 @@ void main({void Function(String) log = print}) {
   }).toList()
     ..removeWhere((element) => !RegExp(r'^[a-zA-Z]*$').hasMatch(element.word));
 
-  final pool = {...wordsList.map((word) => word.word)};
+  final pool = wordsList.map((word) => word.word);
 
   log('Sorting ${pool.length} words');
   final wordPool = WordPool(words: pool);
   log('Sorted ${pool.length} words');
 
-  final generator = AsymmetricalCrosswordGenerator(
+  final generator = SymmetricalCrosswordGenerator(
     pool: wordPool,
     crossword: Crossword(
-      bounds: Bounds.square(size: 78),
+      bounds: Bounds.square(size: 93),
       largestWordLength: wordPool.longestWordLength,
       shortestWordLength: wordPool.shortestWordLength,
     ),
   );
 
+  final stopwatch = Stopwatch()..start();
   final crossword = generator.populate();
+  stopwatch.stop();
 
-  log('Placed words: ${crossword.words.length}');
-  File('asymmetrical_crossword.txt')
+  log('Elapsed time: ${stopwatch.elapsed.inSeconds} seconds');
+  log('Generated a crossword with: ${crossword.words.length}');
+
+  // Creates file to see the crossword
+  File('symmetrical_crossword.txt')
       .writeAsStringSync(crossword.toPrettyString());
 
   // Creates CSV file representing the crossword.
@@ -53,8 +61,7 @@ void main({void Function(String) log = print}) {
 
   final notUsedWords = wordsList.where((w) => !wordsString.contains(w.word));
 
-  log('Unused words: ${notUsedWords.map((w) => w.word).toList()}');
+  log('Unused words: ${notUsedWords.length}');
 
-  File('board_asymmetrical.txt')
-      .writeAsStringSync(const ListToCsvConverter().convert(list));
+  File('board.txt').writeAsStringSync(const ListToCsvConverter().convert(list));
 }
